REF: Refactor DBN structure
