+++ rust/dbn/src/encode.rs
